+++ examples/examples-record.R
@@ -1,0 +1,20 @@
+
+\dontrun{
+
+# use digest 0.6.22 from package repositories -- different ways
+# of specifying the remote. use whichever is most natural
+renv::record("digest@0.6.22")
+renv::record(list(digest = "0.6.22"))
+renv::record(list(digest = "digest@0.6.22"))
+
+# alternatively, provide a full record as a list
+digest_record <- list(
+  Package = "digest",
+  Version = "0.6.22",
+  Source  = "Repository",
+  Repository = "CRAN"
+)
+
+renv::record(list(digest = digest_record))
+
+}